refactor: floating button bottom값 조정
+ fix: tab을 searchParams로 관리하다보니 생기는 문제 해결
tab을 바꾼 후 뒤로가기 버튼을 눌렀을 때 기대되는 것은 해당 화면에서 나가는 것
=> 실제로는 tab이 이전 tab으로 바뀌는 식으로 동작
