+++ src/app/(routes)/profile/star-history/layout.tsx
@@ -1,8 +1,7 @@
 import { FunctionComponent, PropsWithChildren } from 'react'
 import type { Metadata } from 'next'
 import BottomNavLayout from '@/shared/components/custom/bottom-nav-layout'
-import GoBackButton from '@/shared/components/custom/go-back-button'
-import Text from '@/shared/components/ui/text'
+import Header from './@header/default'
 
 export const metadata: Metadata = {}
 
@@ -11,12 +10,7 @@ interface LayoutProps extends PropsWithChildren {}
 const Layout: FunctionComponent<LayoutProps> = ({ children }) => {
   return (
     <BottomNavLayout where="마이">
-      <header className="relative flex h-[54px] w-full items-center bg-background-base-01 px-[16px]">
-        <GoBackButton />
-        <Text typography="subtitle2-medium" className="center">
-          별 내역
-        </Text>
-      </header>
+      <Header />
 
       {children}
     </BottomNavLayout>
